Split out formatProduct
